Don't remember what this commit was about

+++ SerenityGardenTD/Assets/_SerenityGardenTD/Scripts/BattleScene/Interfaces/IAttacker.cs
@@ -7,7 +7,7 @@ namespace SerenityGarden
     public interface IAttacker<T>
     {
         T Target { get; set; }
-        int Damage { get; set; }
+        float Damage { get; set; }
         float Range { get; set; }
 
         float AttackCooldown { get; set; }
