+++ SerenityGardenTD/Assets/_SerenityGardenTD/Scripts/BattleScene/Interfaces/IDestroyable.cs
@@ -6,7 +6,7 @@ namespace SerenityGarden
 {
     public interface IDestroyable
     {
-        int Health { get; set; }
+        float Health { get; set; }
         int DestroyReward { get; set; }
 
         void Die();
